:recycle: 인증 페이지 레이아웃 tailwind-variants 적용 #17
